integrations/adk-middleware/python/pyproject.toml
[project]
name = "ag_ui_adk"
version = "0.3.6"
readme = "README.md"
authors = [
    { name = "Mark Fogle", email = "mark@contextable.com" }
]
requires-python = ">=3.9, <3.14"
dependencies = [
    "ag-ui-protocol>=0.1.10",
    "aiohttp>=3.12.0",
    "asyncio>=3.4.3",
    "fastapi>=0.115.2",
    "google-adk>=1.14.0",
    "pydantic>=2.11.7",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "pluggy>=1.6.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
]
